Cleanup utils interface, create subpackage
